検索ボックスのコンポーネント化
